requirements.txt
python-telegram-bot==20.3
Flask==2.3.3
APScheduler==3.10.1
pytz==2023.3